py_sample/fastapi_custom_logging/pyproject.toml
[tool.poetry]
name = "fastapi_custom_logging"
version = "0.1.0"
description = ""
authors = ["wararaki <ky7.ott.w@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
uvicorn = "^0.11.5"
fastapi = "^0.54.1"
PyYAML = "^5.3"

[tool.poetry.dev-dependencies]
pytest = "^3.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
